pyproject.toml
[project]
name = "new-project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "langchain>=1.0.7",
    "langchain-core>=1.0.5",
    "langchain-openai>=1.0.3",
    "langgraph>=1.0.3",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.32.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
]
